1. added gdb scripts and target description files in support of rael mode debugging using gdb
2. succesfully read the 32k bootloader kernel image correctly from the disk. tested it using a dumy tester file - test.S whose size after being compiled equals the bootloader image size (32k - 512b) and it is filled up with nop instructions except for the very last instruction being "jmp ." the program hangs there expectedly

3. successfully entered protected mode

note. gdtr has to be filled with the *linear* address of the gdt base. i used to give it the address that's relative to the  address zero and it was awfully wrong and took me hours to find out :(
the comment following "gdtr:" should explain it fully and demostrate the correct way to handle it