Unstable version. It works, but it's probably not good.
ide.c is now able to detect all drives across two channels.

I tested it by modifying the QEMU configuration in the Makefile. I added three more VHDs to make a total of four drives, which is as many as IDE supports (two channels, two drives each).

I didn't figure out why insl() didn't work for reading. I changed it to reading one byte at a time and checking the controller status in between, and that worked. Does it only work with interrupts enabled? Is it a blocking instruction? (I guess it is.)