Read the newly added ibm 5170 fixed disk manual about read and write commands to find out the answer to the questions in the previous commit:
After the issuing of a write command, the first sector (in a series of sector reads issues as one read command) can be immediately written into the data io port. And we are confining our ide_abs_rw() to handle one sector at a time. Thus, using outsl() is fine.

After the issuing of a read command, the controller may *not* be ready to accept any daat from the data port. So we *must* wait until the ready bit set. And once the ready bit is set, it indicates a *whole* sector is ready to be read. Thus, to use insl() in a read operation synchronously, we must poll the ready bit first before insl() can be called. Or we may choose to have the controller generate IRQ14 upon which we read!

Look closely into the manual to see when exactly the interrupts are generated.